+++ e3nn/util/codegen/__init__.py
@@ -1,8 +1,6 @@
-from ._eval import eval_code
 from ._mixin import CodeGenMixin
 
 
 __all__ = [
-    "eval_code",
     "CodeGenMixin",
 ]